+++ app/javascript/vue/tasks/observation_matrices/matrix_row_coder/store/mutations/setDescriptorUnsaved.js
@@ -1,0 +1,7 @@
+export default (state, args) => {
+  const {
+    descriptorId,
+    isUnsaved
+  } = args
+  state.descriptors.find(d => d.id === descriptorId).isUnsaved = !!isUnsaved
+}